// src/components/Interfaces/Podcast/IPodcastHostPage.tsx
export interface IPodcastHostPage {
    id: number;
    podcastHostPicture: string;
    podcastHostName: string;
    podcastHostTitle: string;
    podcastHostNickname: string;
    podcastHostFavoriteBands: string;
    podcastHostBio: string;
    podcastHostInstagram: string;
    podcastHostFacebook: string;
    podcastHostLinkedIn: string;
    podcastHostEmailContact: string;
    displayOrder: number;
}